Fix "cannot read text" bug
The error was "Cannot read properties of undefined (reading 'text')".
It happend because it couldn't read the text from a script that didn't
have text.

The problem was that the DOM had many scripts, probably inserted in
different order and I always expected to find the screenName in the
first script of the list.

This commit changes it, so that we look through all scripts for one that
starts with "window.__INITIAL_STATE__"), which is how the text of the
script starts where we find the screenName. If we don't find such a
script we just set the screenName to not found.

Note that querySelectorAll does not return an Array, but a NodeList,
which is why we have to typeCast it into an Array before we can call the
filter() function. NodeList would support a forEach() function, but that
seemed more hacky.